moving from g5 to g1
